Moved the function MainPokemon into collection_dialog.py because it's the only place that needs it
